updated mentionable exmaple values
